+++ polardb-20170801/alibabacloud_polardb20170801/models/_describe_application_performance_response_body.py
@@ -22,14 +22,14 @@ class DescribeApplicationPerformanceResponseBody(DaraModel):
         self.application_id = application_id
         # The application type.
         self.application_type = application_type
-        # The end time of the query. The time is in UTC and follows the `yyyy-MM-ddTHH:mm:ssZ` format.
+        # The end of the time range to query. The time is in the `yyyy-MM-ddTHH:mm:ssZ` format (UTC).
         self.end_time = end_time
+        # The data granularity.
         self.interval = interval
-        # The performance metrics.
         self.performance_keys = performance_keys
         # The request ID.
         self.request_id = request_id
-        # The start time of the query. The time is in UTC and follows the `yyyy-MM-ddTHH:mm:ssZ` format.
+        # The beginning of the time range to query. The time is in the `yyyy-MM-ddTHH:mm:ssZ` format (UTC).
         self.start_time = start_time
 
     def validate(self):